[feat#4] return Box 테이블 추가

+++ src/main/java/carbonneutral/academy/domain/cafe/repository/CafeRepository.java
@@ -3,5 +3,5 @@ package carbonneutral.academy.domain.cafe.repository;
 import carbonneutral.academy.domain.cafe.Cafe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CafeRepository extends JpaRepository<Cafe, Long> {
+public interface CafeRepository extends JpaRepository<Cafe, Integer> {
 }
